pyproject.toml
[project]
name = "apsit-notifier"
version = "0.1.0"
description = "Notification on Whtasapp and Telegram"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.9.5",
    "beautifulsoup4==4.12.3",
    "motor==3.3.2",
    "pymongo==4.6.3",
    "python-dotenv==1.0.1",
    "python-telegram-bot==20.8",
]
